Add support for the Spectral linter and more generalized linting.
